+++ SkyStone-5.3/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Blue_Foundation.java
@@ -2,12 +2,14 @@ package org.firstinspires.ftc.teamcode;
 
 //Importing required classes
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 // Declaring autonomous named Servo_Autonomous with the ground test
+@Disabled
 @Autonomous(name="Blue Foundation", group="Blue")
 // Creating class named servo autonomous that uses linear op mode
 public class Blue_Foundation extends LinearOpMode {
