+++ SkyStone-5.3/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Carte_Blanche_Red.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode;
+
+public class Carte_Blanche_Red {
+}
